Changes of PR1 adressed
